backend/requirements.txt
# Core dependencies
fastapi==0.115.10
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic[email]==2.4.2
pydantic-settings==2.0.3

# Auth and Security
supabase==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Machine Learning and AI
torch>=2.0.0
numpy==1.24.3
scipy==1.11.3
scikit-learn==1.3.1
transformers==4.30.2
sentence-transformers==2.2.2
huggingface-hub==0.16.4
qdrant-client==1.7.0
google-generativeai==0.8.0
langchain==0.1.0

# Document Processing
python-multipart==0.0.6
python-docx==1.0.1
pypdf2==3.0.1
python-slugify==8.0.1
chardet==5.2.0
